Moved SeriesSeriesTypes from MetaData to SeriesType package
